refactor: extract launcher dashboard
